+++ src/lib.rs
@@ -1,15 +1,14 @@
 //! Ripfuzz - An extremely fast Smart contract fuzzer.
 
 pub use evm::{
-    CallFrame, CallFrameKind, Chain, ChainConfig, CheatcodeConfig, Contract, CoverageId,
-    CoverageReport, CoverageReporter, CoverageUpdate, CoverageWriter, DEFAULT_DEPLOYER,
-    DeployInput, DeployLibraryInput, DeployLibraryOutput, DeployOutput, Evmole, ExecOutput,
+    CallFrame, CallFrameKind, Chain, ChainConfig, CheatcodeConfig, CoverageId, CoverageReport,
+    CoverageReporter, CoverageUpdate, CoverageWriter, DEFAULT_DEPLOYER, DeployInput,
+    DeployLibraryInput, DeployLibraryOutput, DeployOutput, Evmole, ExecOutput,
     ExecutionContractCoverage, ExecutionCoverage, ExecutionTraceWriter, ForkDBConfig,
     MockTransport, RpcStats, SetupInput, SetupOutput, SharedCoverage, StorageChange,
     StorageChangeInfo, StorageType, Trace, TraceContext, TraceDisplay, Transaction,
     TransactionResult,
 };
-pub use foundry::{Artifact, ArtifactId, BuildInfo, BuildOptions, Project};
 pub use tester::{
     BrokenInvariant, BrokenInvariantReporter, Corpus as TestCorpus, Fuzzer as TestFuzzer,
     Shrinker as TestShrinker, TestHarness,
@@ -25,4 +24,3 @@ pub mod maxer;
 pub mod tester;
 
 mod evm;
-mod foundry;